+++ tests/Stub/Boolean.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Neat\Configuration\Test\Stub;
+
+use Neat\Configuration\Settings;
+
+class Boolean
+{
+    use Settings;
+
+    const PREFIX = 'TEST_';
+
+    /** @var bool */
+    public $boolean;
+}