+++ tests/Stub/Integer.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Neat\Configuration\Test\Stub;
+
+use Neat\Configuration\Settings;
+
+class Integer
+{
+    use Settings;
+
+    const PREFIX = 'TEST_';
+
+    /** @var int */
+    public $integer;
+}