+++ tests/Stub/Name.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Neat\Configuration\Test\Stub;
+
+use Neat\Configuration\Settings;
+
+class Name
+{
+    use Settings;
+
+    const PREFIX = 'TEST_';
+
+    /** @var string */
+    public $name;
+}