+++ tests/Stub/SkipStatic.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Neat\Configuration\Test\Stub;
+
+use Neat\Configuration\Settings;
+
+class SkipStatic
+{
+    use Settings;
+
+    /** @var string */
+    public static $static;
+}