+++ tests/Stub/Unknown.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Neat\Configuration\Test\Stub;
+
+use Neat\Configuration\Settings;
+
+class Unknown
+{
+    use Settings;
+
+    /** @var string */
+    public $unknown;
+}